fix types, bash the tooling
i SWEAR i hate tooling. The `sed` in the build script is needed because
since I use `import type`, tsc figures that `exporter.ts` is a module
(even though types are supposed to be "erased at compile time"), so it
includes some annoying shim one-liner that tries to define a module,
except that breaks in browser because it's not actually a module.

I need to use `import type` because I couldn't figure out any other way
to have tsc figure out the type of z.infer while keeping the other
globals defined.

All in all this works well enough, if you're a tooling genius god please
fix this, otherwise I give up this is good enough. If you're
developping, you'll just need to not copy/paste the first two lines of
the built exporter.js file.